Moves DirectRecordAccess+Set to .internal
