+++ community/kernel/src/test/java/org/neo4j/kernel/impl/transaction/state/PropertyCreatorTest.java
@@ -32,7 +32,7 @@ import org.neo4j.kernel.impl.store.record.PropertyRecord;
 import org.neo4j.kernel.impl.store.record.Record;
 import org.neo4j.kernel.impl.transaction.state.RecordAccess.Loader;
 import org.neo4j.kernel.impl.transaction.state.RecordAccess.RecordProxy;
-import org.neo4j.unsafe.batchinsert.DirectRecordAccess;
+import org.neo4j.unsafe.batchinsert.internal.DirectRecordAccess;
 import org.neo4j.unsafe.impl.batchimport.store.BatchingIdSequence;
 import org.neo4j.values.storable.Value;
 import org.neo4j.values.storable.Values;
