+++ community/kernel/src/test/java/org/neo4j/unsafe/batchinsert/internal/BatchedFlushStrategyTest.java
@@ -22,8 +22,6 @@ package org.neo4j.unsafe.batchinsert.internal;
 import org.junit.Test;
 import org.mockito.Mockito;
 
-import org.neo4j.unsafe.batchinsert.DirectRecordAccessSet;
-
 public class BatchedFlushStrategyTest
 {
